requirements.txt
Django==5.1.2
channels==3.0.5
channels-redis==3.3.
djangorestframework==3.14.0
